requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.3
Django==5.1
django-cors-headers==4.3.1
django-filter==23.5
django-graphql-jwt==0.4.0
graphene==3.4.3
graphene-django==3.2.0
graphql-core==3.2.6
graphql-relay==3.2.0
idna==3.10
pillow==11.3.0
promise==2.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
requests==2.31.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.15.0
urllib3==2.5.0
